test: replace jest with built-in runner
