+++ backend/package.json
@@ -4,13 +4,14 @@
   "main": "server.js",
   "type": "commonjs",
   "scripts": {
-    "start": "node server.js"
+    "start": "node server.js",
+    "test": "node --test"
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
-    "multer": "^1.4.5-lts.1",
     "js-yaml": "^4.1.0",
-    "cors": "^2.8.5"
+    "multer": "^1.4.5-lts.1"
   }
 }
